+++ app/views/schools/participants/show.html.erb
@@ -93,7 +93,7 @@
             Status
           </dt>
           <dd class="govuk-summary-list__value">
-            <%= render StatusTags::SchoolParticipantStatusTag.new(participant_profile: @profile) %>
+            <%= render StatusTags::SchoolParticipantStatusTag.new(participant_profile: @profile, school: @school) %>
           </dd>
           <dd class="govuk-summary-list__actions"></dd>
         </div>
